requirements.txt
streamlit
mysql-connector-python
transformers
torch
pandas
matplotlib
